fix system.json structure
